requirements.txt
streamlit
pytesseract
pillow
pdf2image
pandas
google-generativeai
